requirements.txt
python-telegram-bot==22.6
google-genai
asyncpg
apscheduler==3.10.4
httpx
python-dotenv
pytz
ruff
python-dateutil
feedparser
